requirements.txt
gradio==3.23.0
torch==2.0.0
numpy==1.24.3
open3d==0.13.0
trimesh==3.21.5
transformers==4.28.1
azure-storage-blob==12.16.0
psutil==5.9.5
pynvml